feat: implement notification system, product management, and order service controllers and models
