+++ controllers/orderController.js
@@ -377,266 +377,19 @@ exports.vnpayIpn = async (req, res) => {
 
 exports.checkout = async (req, res) => {
   try {
-    const userId = req.user.id;
-    const { name, addressReceive, phone, totalPrice, paymentMethod, voucherCode, items } = req.body;
-
-    // validate input - name is the recipient's name (who will receive the order)
-    if (!name || !addressReceive || !phone || !totalPrice || !paymentMethod || !items || !Array.isArray(items) || items.length === 0) {
-      return res.status(400).json({ success: false, message: 'Missing required fields: name, addressReceive, phone, totalPrice, paymentMethod, items' });
-    }
-    if (!['COD', 'VNPAY'].includes(paymentMethod)) {
-      return res.status(400).json({ success: false, message: 'Invalid payment method' });
-    }
-
-    // Check if account exists
-    const account = await Accounts.findById(userId);
-    if (!account) {
-      return res.status(404).json({ success: false, message: 'Account not found' });
-    }
-
-    // Calculate voucher (if any)
-    let voucher = null;
-    let discountAmount = 0;
-    let finalPrice = totalPrice;
-
-    if (voucherCode) {
-      try {
-        const result = await applyVoucher(voucherCode, totalPrice);
-        // applyVoucher returns { success, message, data: { voucher, discountAmount, finalPrice } }
-        if (result.success && result.data) {
-          voucher = result.data.voucher;
-          discountAmount = result.data.discountAmount;
-          finalPrice = result.data.finalPrice;
-        }
-        // If not success (invalid voucher), ignore voucher and keep original price
-      } catch (err) {
-        // Ignore voucher, keep original price
-      }
-    }
-
-    // Create order
-    const newOrder = new Orders({
-      accountId: userId,
-      name,
-      addressReceive,
-      phone,
-      totalPrice,
-      voucherId: voucher ? voucher._id : null,
-      discountAmount,
-      finalPrice,
-      orderStatus: 'pending',
-      payStatus: 'unpaid',
-      paymentMethod,
-    });
-
-    const savedOrder = await newOrder.save();
-
-    // Increment usedCount if voucher was applied
-    if (voucher) {
-      voucher.usedCount += 1;
-      await voucher.save();
-    }
-
-    // Create order details from items
-    const orderDetailsToSave = [];
-    const boughtVariantIds = [];
-    for (const item of items) {
-      const { variantId, unitPrice, Quantity, feedback_details } = item;
-
-      // validate item
-      if (!variantId || !unitPrice || !Quantity) {
-        return res.status(400).json({ success: false, message: 'Invalid item in order details' });
-      }
-      if (unitPrice < 0) {
-        return res.status(400).json({ success: false, message: 'Unit price cannot be negative' });
-      }
-      if (Quantity < 1) {
-        return res.status(400).json({ success: false, message: 'Quantity must be at least 1' });
-      }
-      if (feedback_details && feedback_details.length > 500) {
-        return res.status(400).json({ success: false, message: 'Feedback cannot exceed 500 characters' });
-      }
-
-      // Atomically check-and-deduct stock in a single operation.
-      // If another concurrent checkout already took the last unit, this returns null
-      // and we abort with a clear error instead of overselling.
-      const updatedVariant = await ProductVariant.findOneAndUpdate(
-        { _id: variantId, stockQuantity: { $gte: Quantity } },
-        [
-          {
-            $set: {
-              stockQuantity: { $subtract: ['$stockQuantity', Quantity] },
-              // Auto-deactivate when stock hits 0
-              variantStatus: {
-                $cond: [
-                  { $eq: [{ $subtract: ['$stockQuantity', Quantity] }, 0] },
-                  'inactive',
-                  '$variantStatus',
-                ],
-              },
-            },
-          },
-        ],
-        { new: true }
-      );
-
-      if (!updatedVariant) {
-        // Roll back the order and voucher usage if stock deduction failed
-        await savedOrder.deleteOne();
-        if (voucher) {
-          voucher.usedCount -= 1;
-          await voucher.save();
-        }
-        return res.status(400).json({
-          success: false,
-          message: `Insufficient stock for variant ${variantId}. The item may have just sold out.`,
-        });
-      }
-
-      // Stock confirmed and deducted — create the order detail record
-      const orderDetail = new OrderDetails({
-        orderId: savedOrder._id,
-        variantId,
-        unitPrice,
-        Quantity,
-      });
-      const savedDetail = await orderDetail.save();
-      orderDetailsToSave.push(savedDetail);
-      boughtVariantIds.push(variantId.toString());
-    } // end for (const item of items)
-
-    // Link order details to the order
-    const orderDetailsIds = orderDetailsToSave.map(detail => detail._id);
-    savedOrder.orderDetails = orderDetailsIds;
-    await savedOrder.save();
-
-    // Remove purchased items from the cart
-    const objectUserId = mongoose.Types.ObjectId.isValid(userId) ? new mongoose.Types.ObjectId(userId) : userId;
-    const objectVariantIds = boughtVariantIds.map(id => mongoose.Types.ObjectId.isValid(id) ? new mongoose.Types.ObjectId(id) : id);
-    await Cart.deleteMany({
-      accountId: objectUserId,
-      variantId: { $in: objectVariantIds },
-    });
-
-    // Emit Socket.IO events for cart update and new order
     const io = req.app.get('io');
-    if (io && userId) {
-      // Emit cart update
-      io.to(`user_${userId.toString()}`).emit('cartUpdated', {
-        action: 'cleared',
-        accountId: userId
-      });
-
-      // Populate order with user details for Socket.IO emission
-      const populatedOrder = await Orders.findById(savedOrder._id)
-        .populate('accountId', 'username name email phone')
-        .lean();
-
-      // Emit new order creation for real-time updates with populated data
-      const formattedOrderForSocket = {
-        _id: populatedOrder._id,
-        accountId: populatedOrder.accountId,
-        name: populatedOrder.name,
-        addressReceive: populatedOrder.addressReceive,
-        phone: populatedOrder.phone,
-        totalPrice: populatedOrder.totalPrice,
-        voucherId: populatedOrder.voucherId,
-        discountAmount: populatedOrder.discountAmount,
-        finalPrice: populatedOrder.finalPrice,
-        orderStatus: populatedOrder.orderStatus,
-        payStatus: populatedOrder.payStatus,
-        paymentMethod: populatedOrder.paymentMethod,
-        orderDate: populatedOrder.orderDate,
-        createdAt: populatedOrder.createdAt,
-        updatedAt: populatedOrder.updatedAt || populatedOrder.createdAt,
-        orderDetails: orderDetailsIds
-      };
-
-      io.to(`user_${userId.toString()}`).emit('orderUpdated', {
-        userId: userId.toString(),
-        order: formattedOrderForSocket
-      });
-      // Also notify admins
-      io.to('order_admins').emit('orderUpdated', {
-        userId: userId.toString(),
-        order: formattedOrderForSocket
-      });
-
-      console.log(`New order ${savedOrder._id} created, emitted to user_${userId} and order_admins`);
-
-      // Create and emit order creation notification
-      try {
-        const notification = await createOrderNotification({
-          userId: userId.toString(),
-          orderId: savedOrder._id.toString(),
-          orderStatus: savedOrder.orderStatus,
-          payStatus: savedOrder.payStatus,
-          messageType: 'created'
-        });
-
-        // Emit notification immediately
-        emitOrderNotification(io, notification, userId.toString());
-      } catch (notifError) {
-        console.error('Error creating order creation notification:', notifError);
-      }
-    }
-
+    const result = await orderService.checkoutService(req.user.id, req.body, io);
+    
     return res.status(201).json({
       success: true,
       message: 'Order created successfully with details, cart cleared',
-      data: {
-        order: {
-          _id: savedOrder._id,
-          accountId: savedOrder.accountId,
-          addressReceive: savedOrder.addressReceive,
-          phone: savedOrder.phone,
-          totalPrice: savedOrder.totalPrice,
-          voucherId: savedOrder.voucherId,
-          discountAmount: savedOrder.discountAmount,
-          finalPrice: savedOrder.finalPrice,
-          orderStatus: savedOrder.orderStatus,
-          payStatus: savedOrder.payStatus,
-          paymentMethod: savedOrder.paymentMethod,
-          orderDate: savedOrder.orderDate,
-          orderDetails: orderDetailsIds
-        },
-        orderDetails: orderDetailsToSave.map(detail => ({
-          _id: detail._id,
-          orderId: detail.orderId,
-          variantId: detail.variantId,
-          unitPrice: detail.unitPrice,
-          Quantity: detail.Quantity,
-          feedback: detail.feedback
-        })),
-        orderDetailsIds: orderDetailsIds,
-        voucher: voucher ? {
-          _id: voucher._id,
-          code: voucher.code,
-          discountType: voucher.discountType,
-          discountValue: voucher.discountValue,
-          minOrderValue: voucher.minOrderValue,
-          maxDiscount: voucher.maxDiscount,
-          usedCount: voucher.usedCount,
-          usageLimit: voucher.usageLimit,
-          startDate: voucher.startDate,
-          endDate: voucher.endDate,
-          isDeleted: voucher.isDeleted
-        } : null,
-        summary: {
-          totalItems: orderDetailsToSave.length,
-          totalQuantity: orderDetailsToSave.reduce((sum, detail) => sum + detail.Quantity, 0),
-          originalPrice: savedOrder.totalPrice,
-          discountAmount: savedOrder.discountAmount,
-          finalPrice: savedOrder.finalPrice,
-          stockDeducted: orderDetailsToSave.length
-        }
-      },
+      data: result
     });
-
   } catch (error) {
-    return res.status(500).json({
-      success: false,
-      message: error.message || 'Internal server error',
+    console.error('Checkout error:', error);
+    res.status(error.status || 500).json({ 
+      success: false, 
+      message: error.message || 'Error creating order' 
     });
   }
 };
